requirements.txt
asgiref==3.8.1
# 
crispy-bootstrap5==2024.2
# Библиотека для совместимости Django с Bootstrap5, упрощает создания форм с использованием Bootstrap5
Django==5.0.6
# Фреймворк для создания веб-приложений
django-crispy-forms==2.1
# Приложение Джанго которое помогает создавать красивые формы
sqlparse==0.5.0
# парсер для работы с базами данных SQL
Pillow==10.3.0 
# Библиотека для работы с изображениями
django-colorfield==0.11.0 
# Библиотека которая используется для добавления "выбора цвета или цветовых полей" в моделях джанго
django-filter==24.2
# Библиотека которая позволяет легко создавать фильтры для данных в приложении
mysqlclient==2.2.4
# Джанго использует для подключения и работ с базами данных
faker==25.8.0 
# генерация потдельных данных 
python-dateutil==2.9.0.post0 
# библиотека для работыы с датами и временем
six==1.16.0
# Предназначен для потдержки баз кода, которые работают как на Python 2 так и на Python 3
python-dotenv==1.0.1
""" 
считывает ключ-значение с файла .env и
может устанавливать их как переменные
(У нас в проекте используется для хранения конфиденциальных данных) )

"""
algoliasearch==3.0.0 
# Официальная библиотека клиента Алголия для пайтона, используется для взаимодействия с поисковыми системами Алголии 
algoliasearch==django-3.0.0
# Библиотека для интеграции Алголии с Джанго
certifi==2024.6.2
# 
charset-normalizer==3.3.2
# Библиотека, которая помогает вам читать текст из неизвестной кодировки.
idna==3.7
# Используется для работы с URL-адресами, содержащими не-ASCII символы.
requests==2.32.3 
# Библиотека используемая для выполнения HTTP запросов
urllib3==2.2.1
# Библиотека для работы с HTTP
django-countries==7.6.1
# Библиотека для Джанго которая позволяет использовать список стран в моделях и формах
typing-extensions==4.12.2
# 